requirements.txt
Django<3.4,>3.0
django-elasticsearch-dsl-drf
django-filter
